Fix Poster Coulour, Poll Length and Advanced Editor Bugs
When Clicking on "Advanced Editor", all information was lost, polls, all
checks, like signature check, bbcode check etc.

And the poll_length when used gave it sql error:
"Out of range value for column 'poll_length' at row 1 [1264]"
Because it should be first converted to days, because it is sabe on
database on seconds. ^^

Poster Colour was not used. Now it works.